Improve binary model infrastructure.
Move the share_parameters handling to a subclass of BinaryModel.
That class provides a method which is either the model's name or the list of all
qualified feature names, depending on whether the model shares parameters or not.
